implement BestFirst, pair coding w/ Chris and me
