+++ src/main/java/edu/wpi/aquamarine_axolotls/pathplanning/BestFirstSearch.java
@@ -2,32 +2,29 @@ package edu.wpi.aquamarine_axolotls.pathplanning;
 
 import edu.wpi.aquamarine_axolotls.db.DatabaseController;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class Dijkstra extends AbsHeuristicBased{
+public class BestFirstSearch extends AbsHeuristicBased{
 
-    public Dijkstra(){
+    public BestFirstSearch() {
         super();
     }
 
-    public Dijkstra(List<Node> nodeList, List<Edge> edgeList) {
-        super(nodeList, edgeList);
+    public BestFirstSearch(List<Node> nodeList, List<Edge> edgeList) {
+        super(nodeList,edgeList);
     }
 
     /**
      * Gets the heuristic to use for the priority queue
      * @param next The next node to go to
      * @param goal The goal node to go to
-     * @return 0.0, because Dijkstra doesn't consider the cost of future travel
+     * @return The direct straight-line cost from next to goal
      */
-    double getPriorityHeuristic(Node next, Node goal){
-        return 0.0;
+    double getPriorityHeuristic(Node next, Node goal) {
+        return getCostTo(next, goal);
     }
 
     /**
@@ -35,9 +32,9 @@ public class Dijkstra extends AbsHeuristicBased{
      * @param current The current node in the path
      * @param next The node for which we want the cost
      * @param costSoFar A map of nodes and the costs to get to those nodes
-     * @return The cost to get from the start node to the next node via the current node
+     * @return 0.0, because BestFirst doesn't consider actual cost of travel
      */
     double calculateNewCost(Node current, Node next, Map<Node, Double> costSoFar){
-        return costSoFar.get(current) + getCostTo(current, next);
+        return 0.0;
     }
 }
